src/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.7.0
django-environ==0.10.0
django-filter==23.2
django-js-asset==2.1.0
gevent==23.7.0
greenlet==2.0.2
Pillow==10.0.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing_extensions==4.7.1
zope.event==5.0
zope.interface==6.0
